requirements.txt
gunicorn
setuptools>=42
wheel
Flask==3.1.0
python-dotenv==1.0.1
openai==0.28.0
requests==2.32.3
pandas==2.1.2
numpy==1.26.1
aiohttp==3.11.12
aiosignal==1.3.2
async-timeout==5.0.1
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cryptography==3.4.8
Deprecated==1.2.18
distro==1.7.0
googletrans==4.0.2
groq==0.18.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.3
Jinja2==3.1.5
line-bot-sdk==3.14.5
MarkupSafe==3.0.2
oauthlib==3.2.0
paramiko==2.9.3
PyJWT==2.3.0
pydantic==2.10.6
pydantic_core==2.27.2
PyYAML==6.0.2
rich==13.9.4
six==1.16.0
sniffio==1.3.1
together==1.4.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.3.0
Werkzeug==3.1.3
beautifulsoup4==4.13.3
soupsieve==2.6
spotipy==2.25.0
pydub==0.25.1

# ✅ 新增但不刪除 Render 既有依賴：
aenum==3.1.15
bcrypt==3.2.0
cloudscraper==1.2.71
colorama==0.4.4
fake-useragent==2.0.3
filelock==3.17.0
h2==4.2.0
hyperframe==6.1.0
multidict==6.1.0
packaging==24.1
playwright==1.50.0
pillow==11.1.0
pyee==12.1.1
soupsieve==2.6
tabulate==0.9.0
tqdm==4.67.1
trio==0.28.0
websocket-client==1.8.0
websockets==14.2
yarl==1.18.3
zipp==1.0.0
beautifulsoup4==4.13.3
playwright
playwright-stealth
uvicorn
wikipedia-api==0.8.1
google-generativeai
pillow
supabase

